requirement.txt
Flask==3.0.0
flask-cors==4.0.0
qrcode==7.4.2
Pillow==10.1.0
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pandas==2.1.4
openpyxl==3.1.2